Don't abort build on lint errors
Some setups seem to be enforcing lint checks and preventing build. The extreme number of lint errors found by the lint check is due to a large number of temporarily disabled components that were present in the upstream qualcomm repository, but which made quickly building a user-mode HFP-client very challenging.

Until those components are reintroduced, we will ignore lint errors.

Thanks to @hvdwolf for identifying this glitch and suggesting the workaround.